requirements.txt
fastmcp>=0.1.0
boto3>=1.26.0
pydantic>=2.0.0
python-dotenv>=1.0.0 